Finalize DAO and Service for Beer with new ORM fix beer_test for DAO and Service

+++ apis/beer.go
@@ -11,12 +11,12 @@ import (
 type (
 	// beerService specifies the interface for the beer service needed by beerResource.
 	beerService interface {
-		Get(rs app.RequestScope, id int) (*models.Beer, error)
+		Get(rs app.RequestScope, id uint) (*models.Beer, error)
 		Query(rs app.RequestScope, offset, limit int) ([]models.Beer, error)
 		Count(rs app.RequestScope) (int, error)
 		Create(rs app.RequestScope, model *models.Beer) (*models.Beer, error)
-		Update(rs app.RequestScope, id int, model *models.Beer) (*models.Beer, error)
-		Delete(rs app.RequestScope, id int) (*models.Beer, error)
+		Update(rs app.RequestScope, id uint, model *models.Beer) (*models.Beer, error)
+		Delete(rs app.RequestScope, id uint) (*models.Beer, error)
 	}
 
 	// beerResource defines the handlers for the CRUD APIs.
@@ -41,7 +41,7 @@ func (r *beerResource) get(c *routing.Context) error {
 		return err
 	}
 
-	response, err := r.service.Get(app.GetRequestScope(c), id)
+	response, err := r.service.Get(app.GetRequestScope(c), uint(id))
 	if err != nil {
 		return err
 	}
@@ -85,7 +85,7 @@ func (r *beerResource) update(c *routing.Context) error {
 
 	rs := app.GetRequestScope(c)
 
-	model, err := r.service.Get(rs, id)
+	model, err := r.service.Get(rs, uint(id))
 	if err != nil {
 		return err
 	}
@@ -94,7 +94,7 @@ func (r *beerResource) update(c *routing.Context) error {
 		return err
 	}
 
-	response, err := r.service.Update(rs, id, model)
+	response, err := r.service.Update(rs, uint(id), model)
 	if err != nil {
 		return err
 	}
@@ -108,7 +108,7 @@ func (r *beerResource) delete(c *routing.Context) error {
 		return err
 	}
 
-	response, err := r.service.Delete(app.GetRequestScope(c), id)
+	response, err := r.service.Delete(app.GetRequestScope(c), uint(id))
 	if err != nil {
 		return err
 	}
